% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{check_msconvert_parameter}
\alias{check_msconvert_parameter}
\title{check_msconvert_parameter}
\usage{
check_msconvert_parameter(
  output_format = c("mzXML", "mzML", "mz5", "mgf", "text", "ms1", "cms1", "ms2", "cms2"),
  binary_encoding_precision = c("64", "32"),
  zlib = TRUE,
  write_index = TRUE,
  peak_picking_algorithm = c("vendor", "cwt", "no"),
  vendor_mslevels = c(1, NA),
  cwt_mslevels = c(1, NA),
  cwt_min_snr = 0.1,
  cwt_min_peak_spacing = 0.1,
  subset_polarity = c("any", "positive", "negative"),
  subset_scan_number = c(NA, NA),
  subset_scan_time = c(NA, NA),
  subset_mslevels = c(1, NA),
  zero_samples_mode = c("no", "removeExtra", "addMissing"),
  zero_samples_mslevels = c(1, NA),
  zero_samples_add_missing_flanking_zero_count = 5
)
}
\arguments{
\item{output_format}{"mzXML","mzML", "mz5", "mgf", "text", "ms1",
"cms1", "ms2", or "cms2".}

\item{binary_encoding_precision}{"64" or "32"}

\item{zlib}{TRUE or not.}

\item{write_index}{TRUE or not.}

\item{peak_picking_algorithm}{"vendor" or "cwt"}

\item{vendor_mslevels}{A two numeric vector. Second can be set as NA.}

\item{cwt_mslevels}{A two numeric vector. Second can be set as NA.}

\item{cwt_min_snr}{minimum signal-to-noise ratio}

\item{cwt_min_peak_spacing}{minimum peak spacing}

\item{subset_polarity}{"any", "positive" or "negative"}

\item{subset_scan_number}{A two numeric vector.
Can be c(NA, NA) if don't use this.}

\item{subset_scan_time}{A two numeric vector.
Can be c(NA, NA) if don't use this.}

\item{subset_mslevels}{A two numeric vector. Second can be set as NA.}

\item{zero_samples_mode}{"no", "removeExtra", or "addMissing".}

\item{zero_samples_mslevels}{A two numeric vector. Second can be set as NA.}

\item{zero_samples_add_missing_flanking_zero_count}{= 5}
}
\value{
error or code list
}
\description{
check_msconvert_parameter
}
\examples{
output_format = "mzXML"
binary_encoding_precision = "64"
zlib = TRUE
write_index = TRUE
peak_picking_algorithm = "vendor"
vendor_mslevels = c(1, NA)
cwt_mslevels = c(1, NA)
cwt_min_snr = 0.1
cwt_min_peak_spacing = 0.1
subset_polarity = "any"
subset_scan_number = c(NA, NA)
subset_scan_time = c(NA, NA)
subset_mslevels = c(1, NA)
zero_samples_mode = "no"
zero_samples_mslevels = c(1, NA)
zero_samples_add_missing_flanking_zero_count = 5

check_msconvert_parameter(
  output_format = "mzXML",
  binary_encoding_precision = "64",
  zlib = TRUE,
  write_index = TRUE,
  peak_picking_algorithm = "vendor",
  vendor_mslevels = c(1, 2),
  cwt_mslevels = c(1, NA),
  cwt_min_snr = 0.1,
  cwt_min_peak_spacing = 0.1,
  subset_polarity = "any",
  subset_scan_number = c(NA, NA),
  subset_scan_time = c(NA, NA),
  subset_mslevels = c(1, NA),
  zero_samples_mode = "no",
  zero_samples_mslevels = c(1, NA),
  zero_samples_add_missing_flanking_zero_count = 5
)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
