% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_raw_data.R
\docType{methods}
\name{convert_raw_data}
\alias{convert_raw_data}
\title{convert_raw_data}
\usage{
convert_raw_data(
  input_path = ".",
  output_path,
  msconvert_parameters,
  docker_parameters = c(),
  process_all = FALSE
)
}
\arguments{
\item{input_path}{input path (raw data)}

\item{output_path}{output path. If missing, same with input_path.}

\item{msconvert_parameters}{msconvert parameters. See}

\item{docker_parameters}{docker_parameters}

\item{process_all}{process all raw data together or one by one.}
}
\value{
mzXML or something like this format.
}
\description{
Convert raw data to different format data.
}
\examples{
\dontrun{
msconvert_parameters =
create_msconvert_parameters(
  output_format = "mzXML",
  binary_encoding_precision = "64",
  filter = c("peakPicking true 1-",
             "msLevel 1"),
  zlib = TRUE,
  write_index = TRUE
)
msconvert_parameters
convert_raw_data(msconvert_parameters =
msconvert_parameters)
}
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
