% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msconvert_parameters.R
\docType{methods}
\name{create_msconvert_parameters}
\alias{create_msconvert_parameters}
\title{create_msconvert_parameters}
\usage{
create_msconvert_parameters(
  output_format = c("mzXML", "mzML", "mz5", "mgf", "text", "ms1", "cms1", "ms2",
    "cms2"),
  binary_encoding_precision = c("64", "32"),
  filter = c("peakPicking true 1-", "polarity positive", "polarity negative",
    "msLevel 1", "msLevel 2", "msLevel 3"),
  zlib = TRUE,
  write_index = TRUE
)
}
\arguments{
\item{output_format}{output_format}

\item{binary_encoding_precision}{binary_encoding_precision}

\item{filter}{filter}

\item{zlib}{zlib}

\item{write_index}{write_index}
}
\value{
A msconvert_parameters class object.
}
\description{
Create msconvert parameters for convert_raw_data
}
\examples{
msconvert_parameters =
create_msconvert_parameters(
  output_format = "mzXML",
  binary_encoding_precision = "64",
  filter = c("peakPicking true 1-",
             "msLevel 1"),
  zlib = TRUE,
  write_index = TRUE
)
msconvert_parameters
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
