% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_pull_pwiz.R
\docType{methods}
\name{docker_pull_pwiz}
\alias{docker_pull_pwiz}
\title{docker_pull_pwiz}
\usage{
docker_pull_pwiz(force = FALSE)
}
\arguments{
\item{force}{Force to install pwiz}
}
\value{
docker pwiz image
}
\description{
Pull docker pwiz image
}
\examples{
docker_pull_pwiz(force = FALSE)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
