% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{from_msconvert_parameter_to_code}
\alias{from_msconvert_parameter_to_code}
\title{from_msconvert_parameter_to_code}
\usage{
from_msconvert_parameter_to_code(msconvert_parameter)
}
\arguments{
\item{msconvert_parameter}{msconvert_parameter class object.}
}
\value{
Running code for docker.
}
\description{
From msconvert_parameter class to running code
}
\examples{
parameter1 =
create_msconvert_parameter(
  output_format = "mzXML",
  binary_encoding_precision = "64",
  zlib = TRUE,
  write_index = TRUE,
  peak_picking_algorithm = "vendor",
  vendor_mslevels = c(1, 2),
  cwt_mslevels = c(1, NA),
  cwt_min_snr = 0.1,
  cwt_min_peak_spacing = 0.1,
  subset_polarity = "any",
  subset_scan_number = c(NA, NA),
  subset_scan_time = c(NA, NA),
  subset_mslevels = c(1, NA),
  zero_samples_mode = "no",
  zero_samples_mslevels = c(1, NA),
  zero_samples_add_missing_flanking_zero_count = 5
)
from_msconvert_parameter_to_code(parameter1)

parameter2 =
  create_msconvert_parameter(
    output_format = "mzXML",
    binary_encoding_precision = "32",
    zlib = TRUE,
    write_index = TRUE,
    peak_picking_algorithm = "cwt",
    vendor_mslevels = c(1, NA),
    cwt_mslevels = c(1, NA),
    cwt_min_snr = 0.1,
    cwt_min_peak_spacing = 0.1,
    subset_polarity = "positive",
    subset_scan_number = c(NA, NA),
    subset_scan_time = c(60, 300),
    subset_mslevels = c(1, 2),
    zero_samples_mode = "removeExtra",
    zero_samples_mslevels = c(1, NA),
    zero_samples_add_missing_flanking_zero_count = 5
  )
parameter2
from_msconvert_parameter_to_code(parameter2)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
