% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msconvert_parameters.R
\docType{methods}
\name{from_msconvert_parameters_to_code}
\alias{from_msconvert_parameters_to_code}
\title{from_msconvert_parameters_to_code}
\usage{
from_msconvert_parameters_to_code(msconvert_parameters)
}
\arguments{
\item{msconvert_parameters}{msconvert_parameters class object.}
}
\value{
Running code for docker.
}
\description{
From msconvert_parameters class to running code
}
\examples{
msconvert_parameters =
create_msconvert_parameters(
  output_format = "mzXML",
  binary_encoding_precision = "64",
  filter = c("peakPicking true 1-",
             "msLevel 1"),
  zlib = TRUE,
  write_index = TRUE
)
msconvert_parameters
from_msconvert_parameters_to_code(msconvert_parameters)
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
