% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massconverter_logo.R
\docType{methods}
\name{get_massconverter_version}
\alias{get_massconverter_version}
\title{get_massconverter_version}
\usage{
get_massconverter_version()
}
\value{
version
}
\description{
Get massconverter package version
}
\examples{
get_massconverter_version()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
