% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massconverter_logo.R
\docType{methods}
\name{massconverter_logo}
\alias{massconverter_logo}
\title{massconverter_logo}
\usage{
massconverter_logo()
}
\value{
logo
}
\description{
Get the detailed information of massconverter package.
}
\examples{
massconverter_logo()
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
