% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{massconverter_packages}
\alias{massconverter_packages}
\title{List all packages in the massconverter}
\usage{
massconverter_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include massconverter in the list?}
}
\value{
massconverter_packages
}
\description{
List all packages in the massconverter
}
\examples{
massconverter_packages()
}
