% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msconvert_parameter-class.R
\docType{class}
\name{msconvert_parameter-class}
\alias{msconvert_parameter-class}
\title{An S4 class that stores the parameters}
\description{
An S4 class that stores the parameters
}
\section{Slots}{

\describe{
\item{\code{pacakge_name}}{pacakge_name}

\item{\code{function_name}}{function_name}

\item{\code{parameter}}{parameter}

\item{\code{time}}{time}
}}

