% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msconvert_parameters.R
\docType{class}
\name{msconvert_parameters-class}
\alias{msconvert_parameters-class}
\title{An S4 class that stores the parameters}
\description{
An S4 class that stores the parameters
}
\section{Slots}{

\describe{
\item{\code{output_format}}{output format, mzML, mzXML, mz5, mgf, text, ms1, cms1,
ms2, or cms2.}

\item{\code{binary_encoding_precision}}{binary encoding, 64 or 32}

\item{\code{filter}}{a spectrum list filter}

\item{\code{zlib}}{use zlib compression for binary data or not.}

\item{\code{write_index}}{write index or not.}
}}

