% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_raw_data.R
\docType{methods}
\name{convert_raw_data}
\alias{convert_raw_data}
\title{convert_raw_data}
\usage{
convert_raw_data(
  input_path = ".",
  output_path,
  msconvert_parameter,
  docker_parameters = c(),
  process_all = FALSE
)
}
\arguments{
\item{input_path}{input path (raw data)}

\item{output_path}{output path. If missing, same with input_path.}

\item{msconvert_parameter}{msconvert parameters. See}

\item{docker_parameters}{docker_parameters}

\item{process_all}{process all raw data together or one by one.}
}
\value{
mzXML or something like this format.
}
\description{
Convert raw data to different format data.
}
\examples{
parameter =
  create_msconvert_parameter(
    output_format = "mzXML",
    binary_encoding_precision = "32",
    zlib = TRUE,
    write_index = TRUE,
    peak_picking_algorithm = "cwt",
    vendor_mslevels = c(1, NA),
    cwt_mslevels = c(1, NA),
    cwt_min_snr = 0.1,
    cwt_min_peak_spacing = 0.1,
    subset_polarity = "positive",
    subset_scan_number = c(NA, NA),
    subset_scan_time = c(60, 300),
    subset_mslevels = c(1, 2),
    zero_samples_mode = "removeExtra",
    zero_samples_mslevels = c(1, NA),
    zero_samples_add_missing_flanking_zero_count = 5
  )
parameter

convert_raw_data(input_path = "demo_data/raw_data",
                 output_path = "demo_data/mzxml",
                 msconvert_parameter = parameter,
                 process_all = FALSE
                 )
}
\author{
Xiaotao Shen
\email{shenxt1990@outlook.com}
}
